requirements.txt
numpy==1.24.3
pandas==2.1.4
tensorflow==2.13.0
opencv-python==4.9.0.80
Pillow==10.2.0
scikit-learn==1.4.0
flask==3.0.0
jinja2==3.1.3
werkzeug==3.0.1
python-dotenv==1.0.1
pyserial==3.5
pymongo==4.6.1
certifi==2024.2.2
gunicorn==21.2.0
matplotlib==3.8.2
